cambio float por entero en  ejemplos del paquete1

+++ Ud10_JDBCSesion1/src/paquete1/HolaMundoBaseDatos.java
@@ -29,7 +29,7 @@ public class HolaMundoBaseDatos {
 		
 	while (rs.next()) {
 		String name = rs.getString("titulo");
-		float price = rs.getFloat("precio");
+		int price = rs.getInt("precio");
 		System.out.println(name + "\t" + price);
 	}
 	rs.close();
